@@ -24,4 +24,4 @@ dev_dependencies:
 
 environment:
   sdk: ">=2.12.0-259.9.beta <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
+  flutter: ">=1.20.0"
