@@ -25,7 +25,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
+  MyHomePage({Key? key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -41,7 +41,9 @@ class _MyHomePageState extends State<MyHomePage> {
     final QuickActions quickActions = QuickActions();
     quickActions.initialize((String shortcutType) {
       setState(() {
-        if (shortcutType != null) shortcut = shortcutType;
+        if (shortcutType != null) {
+          shortcut = shortcutType;
+        }
       });
     });
 
@@ -59,7 +61,11 @@ class _MyHomePageState extends State<MyHomePage> {
           type: 'action_two',
           localizedTitle: 'Action two',
           icon: 'ic_launcher'),
-    ]);
+    ]).then((value) {
+      setState(() {
+        shortcut = "actions ready";
+      });
+    });
   }
 
   @override
