@@ -2,14 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(egarciad): Remove once Mockito is migrated to null safety.
-// @dart = 2.9
-
 import 'dart:ui';
 
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
 import 'package:video_player_platform_interface/messages.dart';
 import 'package:video_player_platform_interface/method_channel_video_player.dart';
 import 'package:video_player_platform_interface/test.dart';
@@ -17,13 +13,13 @@ import 'package:video_player_platform_interface/video_player_platform_interface.
 
 class _ApiLogger implements TestHostVideoPlayerApi {
   final List<String> log = [];
-  TextureMessage textureMessage;
-  CreateMessage createMessage;
-  PositionMessage positionMessage;
-  LoopingMessage loopingMessage;
-  VolumeMessage volumeMessage;
-  PlaybackSpeedMessage playbackSpeedMessage;
-  MixWithOthersMessage mixWithOthersMessage;
+  TextureMessage? textureMessage;
+  CreateMessage? createMessage;
+  PositionMessage? positionMessage;
+  LoopingMessage? loopingMessage;
+  VolumeMessage? volumeMessage;
+  PlaybackSpeedMessage? playbackSpeedMessage;
+  MixWithOthersMessage? mixWithOthersMessage;
 
   @override
   TextureMessage create(CreateMessage arg) {
@@ -101,28 +97,11 @@ void main() {
       expect(VideoPlayerPlatform.instance,
           isInstanceOf<MethodChannelVideoPlayer>());
     });
-
-    test('Cannot be implemented with `implements`', () {
-      expect(() {
-        VideoPlayerPlatform.instance = ImplementsVideoPlayerPlatform();
-      }, throwsA(isInstanceOf<AssertionError>()));
-    });
-
-    test('Can be mocked with `implements`', () {
-      final ImplementsVideoPlayerPlatform mock =
-          ImplementsVideoPlayerPlatform();
-      when(mock.isMock).thenReturn(true);
-      VideoPlayerPlatform.instance = mock;
-    });
-
-    test('Can be extended', () {
-      VideoPlayerPlatform.instance = ExtendsVideoPlayerPlatform();
-    });
   });
 
   group('$MethodChannelVideoPlayer', () {
     final MethodChannelVideoPlayer player = MethodChannelVideoPlayer();
-    _ApiLogger log;
+    late _ApiLogger log;
 
     setUp(() {
       log = _ApiLogger();
@@ -140,108 +119,108 @@ void main() {
     test('dispose', () async {
       await player.dispose(1);
       expect(log.log.last, 'dispose');
-      expect(log.textureMessage.textureId, 1);
+      expect(log.textureMessage?.textureId, 1);
     });
 
     test('create with asset', () async {
-      final int textureId = await player.create(DataSource(
+      final int? textureId = await player.create(DataSource(
         sourceType: DataSourceType.asset,
         asset: 'someAsset',
         package: 'somePackage',
       ));
       expect(log.log.last, 'create');
-      expect(log.createMessage.asset, 'someAsset');
-      expect(log.createMessage.packageName, 'somePackage');
+      expect(log.createMessage?.asset, 'someAsset');
+      expect(log.createMessage?.packageName, 'somePackage');
       expect(textureId, 3);
     });
 
     test('create with network', () async {
-      final int textureId = await player.create(DataSource(
+      final int? textureId = await player.create(DataSource(
         sourceType: DataSourceType.network,
         uri: 'someUri',
         formatHint: VideoFormat.dash,
       ));
       expect(log.log.last, 'create');
-      expect(log.createMessage.uri, 'someUri');
-      expect(log.createMessage.formatHint, 'dash');
+      expect(log.createMessage?.uri, 'someUri');
+      expect(log.createMessage?.formatHint, 'dash');
       expect(textureId, 3);
     });
 
     test('create with file', () async {
-      final int textureId = await player.create(DataSource(
+      final int? textureId = await player.create(DataSource(
         sourceType: DataSourceType.file,
         uri: 'someUri',
       ));
       expect(log.log.last, 'create');
-      expect(log.createMessage.uri, 'someUri');
+      expect(log.createMessage?.uri, 'someUri');
       expect(textureId, 3);
     });
 
     test('setLooping', () async {
       await player.setLooping(1, true);
       expect(log.log.last, 'setLooping');
-      expect(log.loopingMessage.textureId, 1);
-      expect(log.loopingMessage.isLooping, true);
+      expect(log.loopingMessage?.textureId, 1);
+      expect(log.loopingMessage?.isLooping, true);
     });
 
     test('play', () async {
       await player.play(1);
       expect(log.log.last, 'play');
-      expect(log.textureMessage.textureId, 1);
+      expect(log.textureMessage?.textureId, 1);
     });
 
     test('pause', () async {
       await player.pause(1);
       expect(log.log.last, 'pause');
-      expect(log.textureMessage.textureId, 1);
+      expect(log.textureMessage?.textureId, 1);
     });
 
     test('setMixWithOthers', () async {
       await player.setMixWithOthers(true);
       expect(log.log.last, 'setMixWithOthers');
-      expect(log.mixWithOthersMessage.mixWithOthers, true);
+      expect(log.mixWithOthersMessage?.mixWithOthers, true);
 
       await player.setMixWithOthers(false);
       expect(log.log.last, 'setMixWithOthers');
-      expect(log.mixWithOthersMessage.mixWithOthers, false);
+      expect(log.mixWithOthersMessage?.mixWithOthers, false);
     });
 
     test('setVolume', () async {
       await player.setVolume(1, 0.7);
       expect(log.log.last, 'setVolume');
-      expect(log.volumeMessage.textureId, 1);
-      expect(log.volumeMessage.volume, 0.7);
+      expect(log.volumeMessage?.textureId, 1);
+      expect(log.volumeMessage?.volume, 0.7);
     });
 
     test('setPlaybackSpeed', () async {
       await player.setPlaybackSpeed(1, 1.5);
       expect(log.log.last, 'setPlaybackSpeed');
-      expect(log.playbackSpeedMessage.textureId, 1);
-      expect(log.playbackSpeedMessage.speed, 1.5);
+      expect(log.playbackSpeedMessage?.textureId, 1);
+      expect(log.playbackSpeedMessage?.speed, 1.5);
     });
 
     test('seekTo', () async {
       await player.seekTo(1, const Duration(milliseconds: 12345));
       expect(log.log.last, 'seekTo');
-      expect(log.positionMessage.textureId, 1);
-      expect(log.positionMessage.position, 12345);
+      expect(log.positionMessage?.textureId, 1);
+      expect(log.positionMessage?.position, 12345);
     });
 
     test('getPosition', () async {
       final Duration position = await player.getPosition(1);
       expect(log.log.last, 'position');
-      expect(log.textureMessage.textureId, 1);
+      expect(log.textureMessage?.textureId, 1);
       expect(position, const Duration(milliseconds: 234));
     });
 
     test('videoEventsFor', () async {
-      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
+      ServicesBinding.instance?.defaultBinaryMessenger.setMockMessageHandler(
         "flutter.io/videoPlayer/videoEvents123",
-        (ByteData message) async {
+        (ByteData? message) async {
           final MethodCall methodCall =
               const StandardMethodCodec().decodeMethodCall(message);
           if (methodCall.method == 'listen') {
-            await ServicesBinding.instance.defaultBinaryMessenger
+            await ServicesBinding.instance?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -251,18 +230,18 @@ void main() {
                       'width': 1920,
                       'height': 1080,
                     }),
-                    (ByteData data) {});
+                    (ByteData? data) {});
 
-            await ServicesBinding.instance.defaultBinaryMessenger
+            await ServicesBinding.instance?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
                         .encodeSuccessEnvelope(<String, dynamic>{
                       'event': 'completed',
                     }),
-                    (ByteData data) {});
+                    (ByteData? data) {});
 
-            await ServicesBinding.instance.defaultBinaryMessenger
+            await ServicesBinding.instance?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -273,25 +252,25 @@ void main() {
                         <int>[1235, 4000],
                       ],
                     }),
-                    (ByteData data) {});
+                    (ByteData? data) {});
 
-            await ServicesBinding.instance.defaultBinaryMessenger
+            await ServicesBinding.instance?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
                         .encodeSuccessEnvelope(<String, dynamic>{
                       'event': 'bufferingStart',
                     }),
-                    (ByteData data) {});
+                    (ByteData? data) {});
 
-            await ServicesBinding.instance.defaultBinaryMessenger
+            await ServicesBinding.instance?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
                         .encodeSuccessEnvelope(<String, dynamic>{
                       'event': 'bufferingEnd',
                     }),
-                    (ByteData data) {});
+                    (ByteData? data) {});
 
             return const StandardMethodCodec().encodeSuccessEnvelope(null);
           } else if (methodCall.method == 'cancel') {
@@ -328,8 +307,3 @@ void main() {
     });
   });
 }
-
-class ImplementsVideoPlayerPlatform extends Mock
-    implements VideoPlayerPlatform {}
-
-class ExtendsVideoPlayerPlatform extends VideoPlayerPlatform {}
