@@ -8,7 +8,7 @@ import 'package:connectivity/connectivity.dart';
 import 'package:connectivity_platform_interface/connectivity_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:mockito/mockito.dart';
+import 'package:test/fake.dart';
 
 const ConnectivityResult kCheckConnectivityResult = ConnectivityResult.wifi;
 const LocationAuthorizationStatus kRequestLocationResult =
@@ -33,7 +33,7 @@ void main() {
   });
 }
 
-class MockConnectivityPlatform extends Mock
+class MockConnectivityPlatform extends Fake
     with MockPlatformInterfaceMixin
     implements ConnectivityPlatform {
   Future<ConnectivityResult> checkConnectivity() async {
