@@ -1,4 +1,4 @@
-// Autogenerated from Pigeon (v0.1.19), do not edit directly.
+// Autogenerated from Pigeon (v0.1.21), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 // ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import
 // @dart = 2.12
@@ -18,7 +18,7 @@ class TextureMessage {
 
   static TextureMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
-    return TextureMessage()..textureId = pigeonMap['textureId'] as int;
+    return TextureMessage()..textureId = pigeonMap['textureId'] as int?;
   }
 }
 
@@ -40,10 +40,10 @@ class CreateMessage {
   static CreateMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return CreateMessage()
-      ..asset = pigeonMap['asset'] as String
-      ..uri = pigeonMap['uri'] as String
-      ..packageName = pigeonMap['packageName'] as String
-      ..formatHint = pigeonMap['formatHint'] as String;
+      ..asset = pigeonMap['asset'] as String?
+      ..uri = pigeonMap['uri'] as String?
+      ..packageName = pigeonMap['packageName'] as String?
+      ..formatHint = pigeonMap['formatHint'] as String?;
   }
 }
 
@@ -61,8 +61,8 @@ class LoopingMessage {
   static LoopingMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return LoopingMessage()
-      ..textureId = pigeonMap['textureId'] as int
-      ..isLooping = pigeonMap['isLooping'] as bool;
+      ..textureId = pigeonMap['textureId'] as int?
+      ..isLooping = pigeonMap['isLooping'] as bool?;
   }
 }
 
@@ -80,8 +80,8 @@ class VolumeMessage {
   static VolumeMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return VolumeMessage()
-      ..textureId = pigeonMap['textureId'] as int
-      ..volume = pigeonMap['volume'] as double;
+      ..textureId = pigeonMap['textureId'] as int?
+      ..volume = pigeonMap['volume'] as double?;
   }
 }
 
@@ -99,8 +99,8 @@ class PlaybackSpeedMessage {
   static PlaybackSpeedMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return PlaybackSpeedMessage()
-      ..textureId = pigeonMap['textureId'] as int
-      ..speed = pigeonMap['speed'] as double;
+      ..textureId = pigeonMap['textureId'] as int?
+      ..speed = pigeonMap['speed'] as double?;
   }
 }
 
@@ -118,8 +118,8 @@ class PositionMessage {
   static PositionMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return PositionMessage()
-      ..textureId = pigeonMap['textureId'] as int
-      ..position = pigeonMap['position'] as int;
+      ..textureId = pigeonMap['textureId'] as int?
+      ..position = pigeonMap['position'] as int?;
   }
 }
 
@@ -135,7 +135,7 @@ class MixWithOthersMessage {
   static MixWithOthersMessage decode(Object message) {
     final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return MixWithOthersMessage()
-      ..mixWithOthers = pigeonMap['mixWithOthers'] as bool;
+      ..mixWithOthers = pigeonMap['mixWithOthers'] as bool?;
   }
 }
 
